remove unnecessary casts

+++ demos/common/cpp/models/src/detection_model_faceboxes.cpp
@@ -104,8 +104,8 @@ void calculateAnchors(std::vector<ModelFaceBoxes::Anchor>& anchors, const std::v
 
     for (auto cy : dense_cy) {
         for (auto cx : dense_cx) {
-            anchors.push_back({ static_cast<float>(cx - 0.5f * skx), static_cast<float>(cy - 0.5f * sky),
-                static_cast<float>( cx + 0.5f * skx), static_cast<float>(cy + 0.5f * sky) });  // left top right bottom
+            anchors.push_back({ cx - 0.5f * skx, cy - 0.5f * sky,
+                 cx + 0.5f * skx, cy + 0.5f * sky });  // left top right bottom
         }
     }
 
@@ -188,7 +188,7 @@ std::vector<int> nms(const std::vector<ModelFaceBoxes::Anchor>& boxes, const std
                     auto overlappingWidth = std::min(boxes[idx1].right, boxes[idx2].right) - std::max(boxes[idx1].left, boxes[idx2].left);
                     auto overlappingHeight = std::min(boxes[idx1].bottom, boxes[idx2].bottom) - std::max(boxes[idx1].top, boxes[idx2].top);
                     auto intersection = overlappingWidth > 0 && overlappingHeight > 0 ? overlappingWidth * overlappingHeight : 0;
-                    auto overlap = static_cast<float>(intersection) / (areas[idx1] + areas[idx2] - intersection);
+                    auto overlap = intersection / (areas[idx1] + areas[idx2] - intersection);
 
                     if (overlap >= thresh) {
                         order[j] = -1;
@@ -243,8 +243,8 @@ std::vector<ModelFaceBoxes::Anchor> filterBBoxes(const InferenceEngine::MemoryBl
         auto predW = exp(dw * variance[1]) * anchors[i].getWidth();
         auto predH = exp(dh * variance[1]) * anchors[i].getHeight();
 
-        bboxes.push_back({ static_cast<float>(predCtrX - 0.5f * predW), static_cast<float>(predCtrY - 0.5f * predH),
-                                     static_cast<float>(predCtrX + 0.5f * predW), static_cast<float>(predCtrY + 0.5f * predH) });
+        bboxes.push_back({ predCtrX - 0.5f * predW, predCtrY - 0.5f * predH,
+                                     predCtrX + 0.5f * predW, predCtrY + 0.5f * predH });
 
     }
 
@@ -274,11 +274,11 @@ std::unique_ptr<ResultBase> ModelFaceBoxes::postprocess(InferenceResult& infResu
     result->objects.reserve(keep.size());
     for (auto i : keep) {
         DetectedObject desc;
-        desc.confidence = static_cast<float>(scores.second[i]);
-        desc.x = static_cast<float>(bboxes[i].left / scaleX);
-        desc.y = static_cast<float>(bboxes[i].top / scaleY);
-        desc.width = static_cast<float>(bboxes[i].getWidth() / scaleX);
-        desc.height = static_cast<float>(bboxes[i].getHeight() / scaleY);
+        desc.confidence = scores.second[i];
+        desc.x = bboxes[i].left / scaleX;
+        desc.y = bboxes[i].top / scaleY;
+        desc.width = bboxes[i].getWidth() / scaleX;
+        desc.height = bboxes[i].getHeight() / scaleY;
         desc.labelID =  0;
         desc.label = labels[0];
